+++ src/common/enums/unit-flags.ts
@@ -33,15 +33,16 @@
 // 		status_flag_hallucination = 0x40000000,
 // 		status_flag_lifetime_expired = 0x80000000,
 
-export enum UnitFlags{
-    Completed = 1,
-    GroundedBuilding= 2,
-    Flying= 4,
-    Burrowed = 0x10,
-    InBunker = 0x20,
-    Loaded= 0x40,
-    Cloaked= 0x200,
-    PassivelyCloaked= 0x800,
-    CanTurn= 0x10000,
-    CanMove= 0x20000,
-  };
+export enum UnitFlags {
+  Completed = 1,
+  GroundedBuilding = 2,
+  Flying = 4,
+  Burrowed = 0x10,
+  InBunker = 0x20,
+  Loaded = 0x40,
+  Cloaked = 0x200,
+  PassivelyCloaked = 0x800,
+  CanTurn = 0x10000,
+  CanMove = 0x20000,
+  Gathering = 0x800000,
+};